requirements.txt
sentence-transformers
faiss-cpu
torch
transformers
fastapi
uvicorn
pandas
openai
python-multipart
colorama==0.4.6
kiwipiepy==0.23.1
kiwipiepy_model==0.23.0
numpy==2.4.4
tqdm==4.67.3